get_stmt_ortho_constraints: only check for obviously unsatisfiable constraints
Commit 5a51262 (get_stmt_ortho_constraints: use isl instead of PolyLib,
Sat Sep 4 14:16:45 2010 -0400) introduced a satisfiability check
on the constructed constraint, in order to switch to the opposite constraint
if the constructed constraint is no good.
The test is fairly expensive, however, and in cases where the opposite
constraint would work, the original constraint should be obviously no good,
so use isl_basic_set_fast_is_empty instead.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>