Fix for maxfuse
Fix for maxfuse. --maxfuse does not guarantee maximal fusion and
in certain cases may end up inserting too many cuts. The upper limit
used before treating it as a failure to find more hyperplanes
is not sufficient in all cases. Changing the limit to a function of
nvar. For maxfuse, if hyperplanes are not found by certain depth,
distribute all SCCs; this fixes it for ccsd_t1_UF. --maxfuse will
have to be written to truly represent maxfuse; it's a weak heuristic geared
for maximal fusion at the moment.

Reported-by: Ajay Panyala
Signed-off-by: Uday Reddy <udayreddy@gmail.com>