Add support for VPATH builds
Makefiles, build scripts and tests currently assume that the build
path and source path are identical. This causes VPATH (out-of-tree)
builds to fail.

Steps to reproduce:

  $ git clone https://github.com/bondhugula/pluto.git --recursive
  $ cd pluto
  $ ./autogen.sh
  $ cd ..
  $ mkdir pluto-build && cd pluto-build
  $ ../pluto/configure --prefix=...
  ...
  ==========================
  Configuring PipLib
  ==========================
  ../pluto/configure: line 12568: cd: piplib/: No such file or directory
  ../pluto/configure: line 12570: ./configure: No such file or directory

  ==========================
  Configuring Openscop
  ==========================
  ../pluto/configure: line 12581: cd: openscop/: No such file or directory
  ../pluto/configure: line 12582: ./configure: No such file or directory

  ==========================
  Configuring Clan
  ==========================
  ../pluto/configure: line 12593: cd: clan/: No such file or directory
  ../pluto/configure: line 12594: ./configure: No such file or directory
  checking for clan/Makefile... no
  configure: error: configure in clan/ failed

This patch sets the correct paths for the configuration and
compilation of submodules and fixes the paths in the main test
script. An update of the reference to candl was also necessary, due to
a minor bug in the candle sources preventing VPATH builds for candle
itself for all versions prior to commit a465102.

Tested for in-tree builds (e.g., ./configure ...), VPATH builds with a
relative path to the sources (e.g., ../pluto/configure ...) and VPATH
builds with an absolute path (e.g., /usr/src/pluto/configure ...).

Signed-off-by: Andi Drebes <andi@drebesium.org>