Fix path specification passed to AC_CHECK_PROG when checking for FileCheck
The set of locations in which `AC_CHECK_PROG` looks for a program is
passed to the macro as a colon-separated path specification.

However, in the invocation of `AC_CHECK_PROG` for `FileCheck`, this
specification is preceded by an extra `path = `, which causes the
first location to be interpreted as `path = $CLANG_PREFIX/bin` rather
than `$CLANG_PREFIX/bin`. This causes the check to fail, even if the
file `$CLANG_PREFIX/bin/FileCheck` exists and is executable.

Removing the prefix makes the check work as expected.

Signed-off-by: Andi Drebes <andi@drebesium.org>